# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Masha Api Documentation",
        "description": "Api docs for Masha Application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "description": "Retrieves all users from the system.",
                "operationId": "da52e0652a921c82fdc0b47e089dab67",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "personal_number": {
                                                "type": "string"
                                            },
                                            "phone_number": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "employee_type": {
                                                "type": "integer"
                                            },
                                            "permission": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "permission_name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "לא נמצאו משתמשים במערכת."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "התרחש בעיית שרתת יש לנסות שוב מאוחר יותר."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{search_string}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Search for users by personal number or full name",
                "description": "Searches for users in the system based on the provided personal number or full name.",
                "operationId": "searchUser",
                "parameters": [
                    {
                        "name": "search_string",
                        "in": "path",
                        "description": "Search string (personal number or full name)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "personal_number": {
                                                    "type": "string"
                                                },
                                                "phone_number": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "employee_type": {
                                                    "type": "integer"
                                                },
                                                "permission": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "code_permission": {
                                                            "type": "integer"
                                                        },
                                                        "permission_name": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "חובה לשלוח מחרוזת לחיפוש."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "לא נמצאו משתמשים במערכת."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "התרחשה בעיה בשרת. אנא נסה שוב מאוחר יותר."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "Creates a new user in the system.",
                "operationId": "storeUser",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "personal_number",
                                    "phone_number",
                                    "employee_type",
                                    "permission_code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "personal_number": {
                                        "type": "string",
                                        "format": "numeric",
                                        "pattern": "^\\d{7}$"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "format": "phone",
                                        "pattern": "^05[0-9]{8}$"
                                    },
                                    "employee_type": {
                                        "type": "integer",
                                        "format": "int32",
                                        "maximum": 4,
                                        "minimum": 1
                                    },
                                    "permission_code": {
                                        "description": "Permission code from the permissions table",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "המשתמש התווסף בהצלחה."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "המשתמש אינו מורשה לבצע פעולה זו."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "משתמש קיים במערכת."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "התרחשה בעיה בשרת. אנא נסה שוב מאוחר יותר."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}